requirements.txt
absl-py==1.4.0
apturl==0.5.2
asttokens==2.4.0
astunparse==1.6.3
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.1
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
comm==0.1.4
command-not-found==0.3
contourpy==1.1.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.7.post1
decorator==5.1.1
defer==1.0.6
deprecation==2.1.0
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fasteners==0.14.1
filelock==3.6.0
flatbuffers==23.5.26
fonttools==4.42.1
future==0.18.2
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
httplib2==0.20.2
idna==3.3
imageio==2.31.4
img2pdf==0.4.4
img2table==1.2.1
importlib-metadata==4.6.4
imutils==0.5.4
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jeepney==0.7.1
JPype1==1.4.1
jupyter_client==8.3.1
jupyter_core==5.3.1
keras==2.13.1
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
libclang==16.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
natsort==8.4.0
nest-asyncio==1.5.7
netifaces==0.11.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.0
olefile==0.46
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.1
paramiko==2.9.3
parso==0.8.3
pdf2image==1.16.3
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==8.5.1
Pillow==10.0.1
platformdirs==2.5.1
polars==0.19.3
pooch==1.7.0
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scipy==1.11.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.2
systemd-python==234
tabula-py==2.8.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
tifffile==2023.9.18
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.6
Werkzeug==2.3.7
wrapt==1.15.0
xdg==5
xkit==0.0.0
XlsxWriter==3.1.7
youtube-dl==2021.12.17
zipp==1.0.0
